NAS-119275: Rename files in tree module
