+++ src/app/modules/ix-tree/components/tree-view/tree-view.component.ts
@@ -2,14 +2,14 @@ import { CdkTree } from '@angular/cdk/tree';
 import {
   ChangeDetectionStrategy, Component, HostBinding, ViewChild, ViewEncapsulation,
 } from '@angular/core';
-import { IxTree } from 'app/modules/ix-tree/components/ix-tree/ix-tree.component';
-import { IxTreeNodeOutletDirective } from 'app/modules/ix-tree/directives/ix-tree-node-outlet.directive';
+import { IxTree } from 'app/modules/ix-tree/components/tree/tree.component';
+import { IxTreeNodeOutletDirective } from 'app/modules/ix-tree/directives/tree-node-outlet.directive';
 
 @Component({
   selector: 'ix-tree-view',
   exportAs: 'ixTreeView',
   template: '<ng-container ixTreeNodeOutlet></ng-container>',
-  styleUrls: ['./ix-tree-view.component.scss'],
+  styleUrls: ['./tree-view.component.scss'],
   // eslint-disable-next-line @angular-eslint/use-component-view-encapsulation
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
