+++ src/app/modules/ix-tree/interfaces/tree-virtual-node-data.interface.ts
