+++ src/app/pages/datasets/components/dataset-management/dataset-management.component.ts
@@ -38,8 +38,8 @@ import { DatasetDetails } from 'app/interfaces/dataset.interface';
 import { EmptyConfig } from 'app/interfaces/empty-config.interface';
 import { Job } from 'app/interfaces/job.interface';
 import { WebsocketError } from 'app/interfaces/websocket-error.interface';
-import { IxTreeFlattener } from 'app/modules/ix-tree/ix-tree-flattener';
 import { IxTreeDataSource } from 'app/modules/ix-tree/tree-datasource';
+import { IxTreeFlattener } from 'app/modules/ix-tree/tree-flattener';
 import { ImportDataComponent } from 'app/pages/datasets/components/import-data/import-data.component';
 import { DatasetTreeStore } from 'app/pages/datasets/store/dataset-store.service';
 import { getTreeBranchToNode } from 'app/pages/datasets/utils/get-tree-branch-to-node.utils';
